% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_court}
\alias{base_court}
\title{An object for shot chart plotting}
\format{
A list with 9 items
}
\usage{
base_court
}
\description{
A list which contains a base court for basketball shot location plotting
}
\keyword{datasets}
