% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_court_old}
\alias{base_court_old}
\title{An object for shot chart plotting}
\format{
A list with 9 items
}
\usage{
base_court_old
}
\description{
A list which contains a base court for basketball shot location plotting
for the old college three point line (20.75 feet all around)
}
\keyword{datasets}
