% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamezone_boxscore.R
\name{gamezone_cbb_boxscore}
\alias{gamezone_cbb_boxscore}
\title{Get GameZone boxscore}
\usage{
gamezone_cbb_boxscore(game_id)
}
\arguments{
\item{game_id}{unique GameZone GameID}
}
\description{
Get GameZone boxscore
}
\examples{
\dontrun{
 gamezone_cbb_boxscore(game_id = 2316023)
}

}
\author{
Jack Lichtenstein
}
