% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{gamezone_mbb_master_schedule}
\alias{gamezone_mbb_master_schedule}
\title{Get GameZone master schedule data by date}
\usage{
gamezone_mbb_master_schedule(date, ranked_games = F)
}
\arguments{
\item{date}{date (of the form "2021-03-01")}

\item{ranked_games}{whether or not to only grab the games where at least
one team is ranked (much faster, since it uses the API)}
}
\description{
Get GameZone master schedule data by date
}
\examples{
\dontrun{
 gamezone_mbb_master_schedule(date = "2021-03-01", ranked_games = T)
}

}
\author{
Jack Lichtenstein
}
