% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_summaries.R
\name{gamezone_mbb_player_stats}
\alias{gamezone_mbb_player_stats}
\title{Get GameZone player statistical summary for a given player}
\usage{
gamezone_mbb_player_stats(player_id)
}
\arguments{
\item{player_id}{playerID found in a play-by-play or box score dataframe returned by
\code{gamezone_mbb_pbp} or \code{gamezone_mbb_boxscore}, respectively}
}
\description{
Get GameZone player statistical summary for a given player
}
\examples{
\dontrun{
 gamezone_mbb_player_stats(player_id = 1061640)
}

}
\author{
Jack Lichtenstein
}
