% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_summaries.R
\name{get_player_season_stats}
\alias{get_player_season_stats}
\title{Get GameZone team statistical summary for a given player}
\usage{
get_player_season_stats(player_id)
}
\arguments{
\item{player_id}{playerID found in a play-by-play dataframe returned by
\code{gamezone_cbb_pbp}}
}
\description{
Get GameZone team statistical summary for a given player
}
\examples{
\dontrun{
 get_player_season_stats(player_id = 1061640)
}

}
\author{
Jack Lichtenstein
}
