% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Get GameZone team schedule data for a given season}
\usage{
get_team_schedule(team, season = "2020-21")
}
\arguments{
\item{team}{team name}

\item{season}{season (of the form "2020-21")}
}
\description{
Get GameZone team schedule data for a given season
}
\examples{
\dontrun{
 get_team_schedule(team = "Duke", season = "2018-19")
}

}
\author{
Jack Lichtenstein
}
