% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_summaries.R
\name{get_team_season_stats}
\alias{get_team_season_stats}
\title{Get GameZone team statistical summary for a given season}
\usage{
get_team_season_stats(team, season = "2020-21")
}
\arguments{
\item{team}{team name}

\item{season}{season (of the form "2020-21")}
}
\description{
Get GameZone team statistical summary for a given season
}
\examples{
\dontrun{
 get_team_season_stats(team = "Duke", season = "2018-19")
}

}
\author{
Jack Lichtenstein
}
