% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbp.R
\name{load_gamezone_pbp}
\alias{load_gamezone_pbp}
\title{Load cleaned gamezoneR play-by-play from the data repository}
\usage{
load_gamezone_pbp(seasons)
}
\arguments{
\item{seasons}{A vector of years (formatted like 2020-21) associated with men's college basketball seasons}
}
\description{
Load cleaned gamezoneR play-by-play from the data repository
}
\examples{
\dontrun{
future::plan("multisession")
load_gamezone_pbp(c("2017-18", "2018-19", "2019-20", "2020-21"))
}

}
\author{
Jack Lichtenstein
}
