% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitutions.R
\name{parse_substitutions}
\alias{parse_substitutions}
\title{Parse GameZone substitutions given a play-by-play dataframe}
\usage{
parse_substitutions(pbp)
}
\arguments{
\item{pbp}{play-by-play dataframe from \code{game_zone_cbb_pbp}}
}
\description{
Parse GameZone substitutions given a play-by-play dataframe
}
\examples{
\dontrun{
 parse_substitutions(pbp = game_zone_cbb_pbp(game_id = 2316023))
}

}
\author{
Jack Lichtenstein
}
